backend/requirements.txt
asgiref==3.11.1
brotli==1.2.0
cffi==2.0.0
colorama==0.4.6
cssselect2==0.9.0
Django==6.0.2
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
factory_boy==3.3.3
Faker==40.4.0
fonttools==4.61.1
inflection==0.5.1
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
psycopg2-binary==2.9.11
pycparser==3.0
pydyf==0.12.1
Pygments==2.19.2
PyJWT==2.11.0
pyphen==0.17.2
pytest==9.0.2
pytest-django==4.12.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.0.0
tzdata==2025.3
uritemplate==4.2.0
weasyprint==68.1
webencodings==0.5.1
zopfli==0.4.1
cryptography==42.0.5
requests==2.32.3
